able to remove cookies and set a preference as a logged out user.
for the time being I've installed react-bootstrap into the project so
that I can get all of the configuration around handling cookies done.
Once completed, I may move the actual modal to the webstore-component-library.

The CookiePreferences component will still live in this app however as
the component library does not need to be aware of cookies and such.