return all wares for the given marketplace
the default api setting is to return 25 wares per page. so, we were only
returning 25 wares total on the browse page. we now are returning "all"
wares by choosing to get 2000 wares per page in the api request. I
didn't see a way to set `page: "all"` or `per_page: "all"`.

if a marketplace has more than 2000 wares, we will need to increase the
number.