adjust indentation in utils/api/configurations.js
the eslint job failed due to "incorrect" indentation in the above file.
it wants the switch statement in `#configureDynamicFormUiSchema` to not
have indentation for the `case` statements. although documentation (e.g.
[mdn docs](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/switch))
show the case statements indented.

the change doesn't interfere with the functionality though, so I'm
changing it so the pipeline passes.