fix the broken browse page specs
the intercept command was not ever working. now, we are expecting to use
the real api calls instead of mocked ones. this is to ensure that our
critical paths of browsing and creating a request are always working,
even when the api is updated.

I also refactored part of the browse page specs into the home page spec,
since that's the flow it was actually testing. also edited the cypress
config to make it more dynamic, which will help with this code existing
in the various client codebases.