properly throw errors from fetcher
although we were catching the error in `#fetcher`, we weren't returning
it as an error. it was being returned to the caller as the `data` value.
this means none of our `isError` checks for GET requests were working
properly.
with this commit, we're throwing the error so that we are handling it
on the view correctly.