specify the environment to use for cypress tests
currently, all cypress tests are failing in ci through github actions
with the error below.
```
Error: [next-auth][error][NO_SECRET]
https://next-auth.js.org/errors#no_secret Please define a `secret` in production. MissingSecret [MissingSecretError]: Please define a `secret` in production.
```

the value of `NEXTAUTH_SECRET` has been set in the staging and production
github environments however. running the job in debug mode shows that the
secret's value is returning as null.
```
```

this commit is an attempt to get the tests to pass in ci by specifying the environment to use for cypress tests. hopefully it'll pick up the secret.