turn off cypress record
the test has been running for over 6 minutes now, but no indication that
anything is happening. I read something about that in a thread earlier
this week, but I don't remember the answer. for now, I'm turning off
the record option in the workflow to see if they will pass.